model/requirements.txt
fastapi
uvicorn
torch
torchvision
ultralytics
pillow
python-multipart
